# nextjs-build/src/frontend/TestFramework.Client.WebAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TestFramework.WebAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Files": {
      "get": {
        "tags": [
          "Files"
        ],
        "operationId": "GetUserFilePage",
        "parameters": [
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserFileDtoPaginationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserFileDtoPaginationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserFileDtoPaginationResult"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Files"
        ],
        "operationId": "PostUserFile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "files": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Files/{id}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "operationId": "DeleteUserFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorCodesResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCodesResult": {
        "type": "object",
        "properties": {
          "ErrorCodes": {
            "type": "array",
            "items": {
              "enum": [
                "IdLessThen1",
                "UIElements_NameMaxLength",
                "UIElements_NameIsEmpty",
                "UIElements_NameAlreadyUsed",
                "UIElements_FindByIsEmpty",
                "UIElements_FindByMaxLength",
                "UITestCases_NameMaxLength",
                "UITestCases_NameIsEmpty",
                "UITestCases_NameAreadyUsed",
                "UITestCases_StartUrlIsEmpty",
                "UITestCases_UIElementNotFound",
                "UITestCases_UserFilesNotFound",
                "UITestCases_StepsAreNotUnique",
                "UIEvents_NameIsEmtpy",
                "UIEvents_NameMaxLength",
                "UITestRuns_TestCasesNotFound",
                "UITestRuns_TestCasesAreEmpty",
                "UIPages_NameIsEmpty",
                "UIPages_NameMaxLength",
                "UITestCaseStatusInfo_StartIsEmpty",
                "UITestCaseStatusInfo_EndIsEmpty",
                "UITestCaseStatusInfo_EndLessThenStart",
                "HealthChecks_NameIsEmpty",
                "HealthChecks_NameMaxLength",
                "HealthChecks_UriIsEmpty",
                "UserFiles_FileNameIsEmpty",
                "UserFiles_FileSizeIsNegative",
                "UserFiles_FileNameMaxLength",
                "PageIndexLessThan0",
                "PageSizeLessThen1",
                "DeletionFailed"
              ],
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUserFileDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "FileName": {
            "type": "string",
            "nullable": true
          },
          "FileSize": {
            "type": "integer",
            "format": "int64"
          },
          "CreatedDateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "GetUserFileDtoPaginationResult": {
        "type": "object",
        "properties": {
          "Count": {
            "type": "integer",
            "format": "int32"
          },
          "Data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetUserFileDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}